requirements.txt
streamlit
torch==2.1.0+cpu
torchvision==0.16.0+cpu
pillow
numpy
